mockup/requirements.txt
Django>=2.2,<3.0
django-crispy-forms==1.7.2

